requirements.txt
streamlit
faiss-cpu
sentence-transformers
scikit-learn
scispacy

